refactor(desktop): 提取 assistant 消息状态机
